Cleanups in code and docs
Updated docs to warn that if 'error' events aren't caught,
an "Unhandled event" exception may be thrown.  Additional
info on TLS options was also added including a link to
the Node documentation.

To external users, Connection.channels is now a frozen
clone rather than the live object.  To facilitate this,
the circular reference from Channel to Connection was
removed.

Internally, all properties of Channel and Connection
are now stored in their respective private stores even
if they're exposed publically.  Now there's no ambiguity
about whether a property is accessed with 'this' or '_(this)'.
Intenally, they should be accessed with '_(this)'.